+++ app/www/test/js/sv-SE.f0r73571n6purp0537051mul473h45h.js
@@ -23,6 +23,7 @@ Soho.Locale.addCulture('sv-SE', {
       month: 'd MMMM',
       year: 'MMMM yyyy',
       timestamp: 'HH:mm:ss',
+      hour: 'HH:mm',
       datetime: 'yyyy-MM-dd HH:mm',
       timezone: 'yyyy-MM-dd HH:mm zz',
       timezoneLong: 'yyyy-MM-dd HH:mm zzzz'