+++ src/components/locale/cultures/ar-SA.js
@@ -22,6 +22,7 @@ Soho.Locale.addCulture('ar-SA', {
         month: 'dd MMMM',
         year: 'MMMM yyyy',
         timestamp: 'h:mm:ss a',
+        hour: 'h:mm a',
         datetime: 'yyyy/MM/dd h:mm a',
         timezone: 'yyyy/MM/dd h:mm a zz',
         timezoneLong: 'yyyy/MM/dd h:mm a zzzz'