+++ src/components/locale/cultures/es-AR.js
@@ -22,6 +22,7 @@ Soho.Locale.addCulture('es-AR', {
       month: 'd de MMMM',
       year: 'MMMM de yyyy',
       timestamp: 'HH:mm:ss',
+      hour: 'HH:mm',
       datetime: 'd/M/yyyy HH:mm',
       timezone: 'd/M/yyyy HH:mm zz',
       timezoneLong: 'd/M/yyyy HH:mm zzzz'