+++ src/components/locale/cultures/hr-HR.js
@@ -22,6 +22,7 @@ Soho.Locale.addCulture('hr-HR', {
       month: 'd. MMMM',
       year: 'MMMM yyyy',
       timestamp: 'HH:mm:ss',
+      hour: 'HH:mm',
       datetime: 'dd. MM. yyyy HH:mm',
       timezone: 'dd. MM. yyyy HH:mm zz',
       timezoneLong: 'dd. MM. yyyy HH:mm zzzz'