+++ src/components/locale/cultures/ja-JP.js
@@ -22,6 +22,7 @@ Soho.Locale.addCulture('ja-JP', {
       month: 'M月 d日',
       year: 'yyyy年 M月',
       timestamp: 'H:mm:ss',
+      hour: 'H:mm',
       datetime: 'yyyy/MM/dd H:mm',
       timezone: 'yyyy/MM/dd H:mm zz',
       timezoneLong: 'yyyy/MM/dd H:mm zzzz'