+++ src/components/locale/cultures/lt-LT.js
@@ -22,6 +22,7 @@ Soho.Locale.addCulture('lt-LT', {
       month: 'MMMM d d.',
       year: 'yyyy m. MMMM',
       timestamp: 'HH:mm:ss',
+      hour: 'HH:mm',
       datetime: 'yyyy-MM-dd HH:mm',
       timezone: 'yyyy-MM-dd HH:mm zz',
       timezoneLong: 'yyyy-MM-dd HH:mm zzzz'