+++ src/components/locale/cultures/ru-RU.js
@@ -21,6 +21,7 @@ Soho.Locale.addCulture('ru-RU', {
       month: 'd MMMM',
       year: 'MMMM yyyy г.',
       timestamp: 'HH:mm:ss',
+      hour: 'HH:mm',
       datetime: 'dd.MM.yyyy HH:mm',
       timezone: 'dd.MM.yyyy HH:mm zz',
       timezoneLong: 'dd.MM.yyyy HH:mm zzzz'